Trying to get location of center of the map view

+++ vulnerable_space/vulnerable_space/Location.swift
@@ -10,8 +10,8 @@ import Parse
 
 class Location : PFObject, PFSubclassing {
     
-    @NSManaged var longitude: Int
-    @NSManaged var latitude: Int
+    @NSManaged var longitude: Double
+    @NSManaged var latitude: Double
     
     override class func initialize() {
         struct Static {
